feat(validation): add input validation to all request DTOs
- Add spring-boot-starter-validation dependency
- Annotate AuthRequest, UserRequest, CategoryRequest, ItemRequest,
  OrderRequest, PaymentRequest, PaymentVerificationRequest with
  @NotBlank, @NotNull, @Email, @Size, @Pattern, @DecimalMin constraints
- Add @Valid to all controller @RequestBody parameters
- Add manual Validator injection for multipart endpoints
  (CategoryController, ItemController)
- Add GlobalExceptionHandler to return structured 400 responses
  on constraint violations instead of stack traces
- Fix role whitelist in UserRequest with @Pattern(ADMIN|USER)
  to prevent ROLE_ADMIN prefix bypass
- Fix role storage in UserServiceImpl: prefix with ROLE_ after validation
- Fix money types: Double → BigDecimal in OrderRequest, OrderItemRequest,
  OrderEntity, OrderItemEntity, OrderResponse

+++ src/main/java/com/saurabh/quickbill/entity/OrderEntity.java
@@ -8,6 +8,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -27,9 +28,9 @@ public class OrderEntity {
     private String orderId;
     private String customerName;
     private String phoneNumber;
-    private Double subTotal;
-    private Double tax;
-    private Double grandTotal;
+    private BigDecimal subTotal;
+    private BigDecimal tax;
+    private BigDecimal grandTotal;
 
     private LocalDateTime createdAt;
